+++ Items/FreezeWorld/FreezeItem.cs
@@ -11,7 +11,7 @@ namespace IssaPlugin.Items
 
         public static void GiveFreezeToLocalPlayer()
         {
-            ItemHelper.GiveItemToLocalPlayer(FreezeItemType, Configuration.FreezeUses.Value, "Freeze");
+            ItemHelper.GiveItemToLocalPlayer(FreezeItemType, (int)Configuration.FreezeUses.Value, "Freeze");
         }
     }
 }
